Remove idea to run bisimulation (in general), at least for version 7.
Bisimulation can't be run for the typical user -- just converting to CompactNFA representation can take tremendous memory, for example in pisot where there are 160K-state NFAs. This is, instead, only run in the optional OTF cases.

This could be revisited if a hook is added to pass  Walnut's internal representation in and out of bisimulation.